+++ Plugins/BigQueryDriverPlugin/BigQueryPlugin.swift
@@ -45,6 +45,7 @@ final class BigQueryPlugin: NSObject, TableProPlugin, DriverPlugin {
     static let supportsSSL = false
     static let tableEntityName = "Tables"
     static let containerEntityName = "Dataset"
+    static let schemaEntityName = "Dataset"
     static let supportsForeignKeyDisable = false
     static let supportsReadOnlyMode = true
     static let databaseGroupingStrategy: GroupingStrategy = .hierarchicalSchema